Fix CSOT unified spec tests.
